// src/interfaces/Knowledge.ts
export interface KnowledgeType {
    _id: string;
    id_: number;
    title: string;
    description: string;
    categorySharing: string;
    fileUrl: string;
    remark: string;
    localCreatedAt: string;
    localUpdatedAt: string;
    isActive: boolean;
    fileTypes: string;
    userCreate: string;
    createdAt: Date;
    updatedAt: Date;
    teamOwner: string;
    imgUrl: string;
}